requirements.txt
asgiref==3.2.10
Django==2.2.16
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-taggit==0.23.0
django-widget-tweaks==1.4.8
Pillow==8.0.0
PyQt5==5.15.0
PyQt5-sip==12.8.1
pytz==2020.1
sqlparse==0.3.1
